requirements.txt
streamlit
opencv-python-headless
mysql-connector-python
numpy
python-dotenv
bcrypt
cryptography